Week 3 Notes.
Some good stuff here regarding actual data science. A nice description of the different kinds of data science that gets done, in terms of difficulty. Some good discussion of what actually constitutes data and big data. And a good description of what a confounding variable is - this was a good reminder for me, because it is always difficult to understand the 'why' behind correlation != causation...well, there can be other 'hidden' variables. Finally, there was a good incorporation of some examples of Experimental design gone wrong, leading to lawsuits, millions of dollars wasted, and people hurt by inappropriate treatments.